updated eval to work both with files and directories
